Revert "gpt -> transformer, template parameterization experiments"
This reverts commit 44d62c3a620513d693e572f0bba77c64f4021521.
